Re-enable skipping over existing analysis files but allow overwriting with `--overwrite`
In a previous commit, the behaviour that the pipeline skips over existing analysis files was removed, but existing analysis files were still checked for. I have re-enabled the behaviour and made it the default. This is so that additional analyses can be easily added without running everything again. Overwriting existing analysis files can be achieved with `--overwrite`.